==> 02 Building and evaluation of ML models.ipynb <==
<center><h1>Building and evaluation of ML models</h1></center>

# 1. Loading preprocessed data set

import pandas as pd

X_tr = pd.read_csv("X_tr.csv")
Y_tr = pd.read_csv("Y_tr.csv", header=None)

print("X_tr.shape: {}\nY_tr.shape: {}".format(X_tr.shape, Y_tr.shape))

import numpy as np

X_values = X_tr.values # taking the numpy matrix of dataframe
Y_values = np.ravel(Y_tr.values) # taking numpy column array from dataframe and converting to simple array

# 2. Creating, fitting and evaluating the models

Methodology used:
* For each model it will be run 10 resamplings, or 10 train/test splits with fitting and evaluation of the model;
* The performance metric adopted for evaluating models in the competition was the f1-score so it will be used here too. As our performance metric is a random variable due to the resampling, it has a mean and standard deviation, making necessary the creation of an objective function with the statistics of the random variable:


$$
\text{maximize} \quad f_o(u) = \mu (u) - 2\sigma(u)
$$

So "best" means higher $f_o(u)$. We can say that the objective function tries to balance the two objectives: 

1. Higher mean, or better generalization;
2. Lower standard deviation, or lower instability.


Obs: Depending on time complexity some hyperparameter optimization was done.

# Auxiliary function to tell when processing is over
def is_over(): # linux os
    import os
    os.system('spd-say "your program has finished"')

# Function for scaling numerical features
from sklearn.preprocessing import MinMaxScaler, StandardScaler

def scale_feat(X_train, X_test, featIndex, scaleType='min-max'):
    if scaleType=='min-max' or scaleType=='std':
        X_tr_norm = np.copy(X_train) # making a copy to let the original available
        X_ts_norm = np.copy(X_test)
        scaler = MinMaxScaler(copy=False) if scaleType=='min-max' else StandardScaler(copy=False)
        scaler.fit(X_tr_norm[:,featIndex])
        X_tr_norm[:,featIndex] = scaler.transform(X_tr_norm[:,featIndex])
        X_ts_norm[:,featIndex] = scaler.transform(X_ts_norm[:,featIndex])
        return (X_tr_norm, X_ts_norm)
    else:
        raise ValueError("Type of scaling not defined. Use 'min-max' or 'std' instead.")

import numpy as np

# Hyperparameters:
# Numerical/Ordinal feautures
numFeat = [
    'building_id', # searching for data leakage
    'vdcmun_id',   # categorical, but used as numerical for simplicity
    'ward_id',     # categorical, but used as numerical for simplicity
    'count_floors_pre_eq',
    'count_floors_post_eq',
    'age_building',
    'plinth_area_sq_ft',
    'height_ft_pre_eq',
    'height_ft_post_eq',
    'count_families'
]

# Train/Test split = 80% train and 20% test
test_size = 0.2

# Index of columns to be scaled
numFeat_idx = np.in1d(X_tr.columns.values, numFeat).nonzero()[0]

# Number of resamplings
n_resamplings = 10

# Objective function
def f_o(u):
    return np.mean(u) - 2*np.std(u)

## 3.1 k-NN

%%time
import datetime
from sklearn.model_selection import train_test_split
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import f1_score

header = ["k", "$\mu$", "$\sigma$", "$f_o$"]
ks = [1, 5, 10] # possible values of k

nn_data = np.zeros((len(ks), len(header)))
for i in range(len(ks)):
    print("Start of k={} at {}".format(ks[i], datetime.datetime.now()))
    results = [0]*n_resamplings
    for j in range(n_resamplings):
        # Train/validation split
        X_train, X_test, y_train, y_test = train_test_split(X_tr.values, Y_tr.values, test_size=test_size)

        # scaling
        X_tr_norm, X_ts_norm = scale_feat(X_train, X_test, featIndex=numFeat_idx, scaleType='min-max')

        # model fitting
        k_nn = KNeighborsClassifier(n_neighbors=ks[i], n_jobs=-1)
        k_nn.fit(X_tr_norm, y_train)

        # model evaluation
        y_pred = k_nn.predict(X_ts_norm)
        results[j] = f1_score(y_test, y_pred, average='weighted')
        
    
    nn_data[i,:] = np.matrix([ks[i], np.mean(results), np.std(results), f_o(results)])

df_knn = pd.DataFrame(nn_data, columns=header)
display(df_knn)
is_over()

# lost OUTPUT:
'''
Start of k=1 at 2019-06-17 00:08:35.281779
Start of k=5 at 2019-06-17 03:17:02.803595
Start of k=10 at 2019-06-17 07:56:13.396357

CPU times: user 3d 11h 17min 47s, sys: 3min, total: 3d 11h 20min 47s
Wall time: 13h 29min 2s
'''

# saving results
# filename = "./simulation_results/df_knn.csv"
# df_knn.to_csv(filename, sep='\t', index=False)

# loading results
df_knn = pd.read_csv("./simulation_results/df_knn.csv", sep='\t')

print("df_knn:")
df_knn

## 3.2 Linear regression

The output was converted to categorical and a linear regression was done. To determine the predicted class the index of the maximum value of the output vector was used.

# Creating the multilabel version of Y_tr
from sklearn.preprocessing import MultiLabelBinarizer

mlb = MultiLabelBinarizer()
Y_tr_multilabel = mlb.fit_transform(Y_tr.values) 

print("Original Y_tr classes:\n{}".format(np.unique(Y_tr)))
print("Caterogical Y_tr classes:\n{}\n".format(np.unique(Y_tr_multilabel,axis=0)))

print("Equivalence:")
for class_ in np.unique(Y_tr_multilabel,axis=0):
    print("{} == {}".format(
        np.argmax(class_)+1,
        class_
    ))

%%time
from sklearn.model_selection import train_test_split
from sklearn import linear_model
from sklearn.metrics import f1_score

scales = ["min-max", "std"]
header = ["scaleType", "$\mu$", "$\sigma$", "$f_o$"]
reg_data = np.empty((len(scales), len(header)), dtype=object)

for i in range(len(scales)):
    results = [0]*n_resamplings
    for j in range(n_resamplings):
        # Train/validation split
        X_train, X_test, y_train, y_test = train_test_split(X_tr.values, Y_tr_multilabel, test_size=test_size)

        # scaling
        X_tr_norm, X_ts_norm = scale_feat(X_train, X_test, featIndex=numFeat_idx, scaleType=scales[i])

        # model fitting
        reg = linear_model.LinearRegression(n_jobs=-1)
        reg.fit(X_tr_norm, y_train)

        # model evaluation
        y_pred = np.argmax(reg.predict(X_ts_norm), axis=1).astype(str) # +1???
        y_pred = mlb.fit_transform(y_pred)

        results[j] = f1_score(y_test, y_pred, average='weighted')


    reg_data[i,:] = np.matrix([scales[i], np.mean(results), np.std(results), f_o(results)])

df_reg = pd.DataFrame(reg_data, columns=header)
is_over()

# saving results
# filename = "./simulation_results/df_reg.csv"
# df_reg.to_csv(filename, sep='\t', index=False)

# loading results
df_reg = pd.read_csv("./simulation_results/df_reg.csv", sep='\t')
df_reg

## 3.3 Logistic regression

%%time
from sklearn import linear_model
from tqdm import tnrange, tqdm_notebook
from sklearn.model_selection import train_test_split
from sklearn.metrics import f1_score

rhos = [0, 0.5, 1] # rho=0 <=> l2 norm / rho==1 <=> l1 norm
Cs = [10**x for x in range(-1,+1 +1)]

header = ["$rho$", "$C$", "$\mu$", "$\sigma$", "$f_o$"]
logreg_data = np.empty((len(rhos)*len(Cs), len(header)), dtype=object)
count=0
for rho in tqdm_notebook(rhos, desc="rho's: "):
    for C in tqdm_notebook(Cs, desc="C's"):
        results = [0]*n_resamplings
        for i in tnrange(n_resamplings, desc='resampling :', leave=False):
            # Train/validation split
            X_train, X_test, y_train, y_test = train_test_split(X_values, Y_values, test_size=test_size)

            # scaling
            X_tr_norm, X_ts_norm = scale_feat(X_train,X_test,featIndex=numFeat_idx,scaleType='min-max')

            # model fitting
            logreg = linear_model.LogisticRegression(multi_class='multinomial',solver='saga',
                                                     penalty='elasticnet',max_iter=500, tol=1e-4, 
                                                     l1_ratio=rho, C=C,  n_jobs=-1)
            logreg.fit(X_tr_norm, y_train)

            # model evaluation
            y_pred = logreg.predict(X_ts_norm)
            results[i] = f1_score(y_test, y_pred, average='weighted')

        logreg_data[count,:] = np.matrix([rho, C, np.mean(results), np.std(results), f_o(results)])
        count+=1

df_logreg = pd.DataFrame(logreg_data, columns=header)
is_over()

# lost OUTPUT:
'''
CPU times: user 15h 57min 10s, sys: 1min 22s, total: 15h 58min 33s
Wall time: 15h 57min
'''

# saving results
# filename = "./simulation_results/df_logreg.csv"
# df_logreg.to_csv(filename, sep='\t', index=False)

# loading results
df_logreg = pd.read_csv("./simulation_results/df_logreg.csv", sep='\t')
print("df_logreg:")
df_logreg

## 3.4 Nearest Centroid Classifier (NCC)

# Creating a class to represent the classifier
from numpy.linalg import norm

class NearestCentroid():
    def __init__(self):
        self.class_list = {}
        self.centroids = {}
    
    def fit(self, X, y):
        self.class_list = np.unique(y, axis=0) # list of classes
        
        # Basically calculates centroids per class
        self.centroids = np.zeros((len(self.class_list), X.shape[1])) # each row is a centroid
        for i in range(len(self.class_list)): # for each class we evaluate its centroid
            temp = np.where(y==self.class_list[i][0])[0]
            self.centroids[i,:] = np.mean(X[temp],axis=0)
            
            
    def predict(self, X):
        N = len(X)
        y_pred = np.zeros(N)
        for i in range(N):
            # find closest
            dist = np.inf
            closest = None
            for j in range(len(self.class_list)):
                temp = X[i] - self.centroids[j]
                new_dist = np.dot(temp,temp)
                if new_dist < dist:
                    dist = new_dist
                    closest = self.class_list[j][0]
            
            y_pred[i] = closest
        
        return y_pred

%%time
scales = ['min-max', 'std']
header = ["scaleType", "$\mu$", "$\sigma$", "$f_o$"]

ncc_data = np.empty((len(scales), len(header)), dtype=object)

count = 0
for scale in scales:
    print("Started scaleType='{}' at {}".format(scale, datetime.datetime.now()))
    results = [0]*n_resamplings
    for i in range(n_resamplings):
        # Train/validation split
        X_train, X_test, y_train, y_test = train_test_split(X_tr.values, Y_tr.values, test_size=test_size)
        
        # scaling
        X_tr_norm, X_ts_norm = scale_feat(X_train, X_test, featIndex=numFeat_idx, scaleType=scale)
        
        # model fitting
        ncc = NearestCentroid()
        ncc.fit(X_tr_norm, y_train)
        
        # model evaluation
        y_pred = ncc.predict(X_ts_norm)
        results[i] = f1_score(y_test, y_pred, average='weighted')
        
    ncc_data[count,:] = np.matrix([scale, np.mean(results), np.std(results), f_o(results)])
    count+=1
    

df_ncc = pd.DataFrame(ncc_data, columns=header)
display(df_ncc)
is_over()

## 3.5 Quadratic Gaussian Classifier (QGC)

During the construction of the *CQG* classifier, was found a problem: the rank of the covariance matrices of the classes is smaller than the number of attributes of the data, which was verified with the following code:

%%time
from numpy.linalg import matrix_rank

# Train/validation split
X_train, X_test, y_train, y_test = train_test_split(X_tr.values, Y_tr.values, test_size=test_size)
X_tr_norm, X_ts_norm = scale_feat(X_train, X_test, featIndex=numFeat_idx, scaleType=scale)

# means evaluation
class_list = np.unique(y_train, axis=0) # list of classes
# centroids per class
centroids = np.zeros((len(class_list), X_tr_norm.shape[1])) # each row is a centroid
for i in range(len(class_list)): # for each class we evaluate its centroid
    temp = np.where(y_train==class_list[i][0])[0]
    centroids[i,:] = np.mean(X_tr_norm[temp],axis=0)

# Covariance matrices evaluation
C_list = [np.zeros((X_tr_norm.shape[1],X_tr_norm.shape[1]))]*len(class_list)
for i in range(len(class_list)):
    # points of class class_list[i]:
    temp = X_tr_norm[np.where(y_train==class_list[i][0])[0]]
    for datapoint in temp:
        aux = np.asmatrix(datapoint - centroids[i])
        C_list[i] += np.matmul(aux.T,aux)
    C_list[i] /= len(temp) # divide by the number of samples


print("Number of features: {}".format(X_tr_norm.shape[1]))
for i in range(len(C_list)):
    print("Rank of C_{} = {}".format(i, matrix_rank(C_list[i])))
    

Therefore, the following methodologies were used to make the matrices invertible:
 
* Variant 1: extract only the main diagonal of the matrices;
* Variant 2: aggregated covariance matrix (*pooled*);
* Variant 3: Friedman regularization, in this case, a grid search was performed to find the best value of hyper-parameter $ \lambda $.

from numpy.linalg import inv
from numpy import dot, matmul

class QuadraticGaussianClassifier():
    def __init__(self, variant=None, Lambda=None):
        self.variant = variant
        self.Lambda = Lambda
        self.class_list = {}
        self.centroids = {}
        self.C_inv = {}
        
    
    def fit(self, X, y):
        self.class_list = np.unique(y, axis=0) # list of classes
        
        # Basically calculates centroids per class
        self.centroids = np.zeros((len(self.class_list), X.shape[1])) # each row is a centroid
        for i in range(len(self.class_list)): # for each class we evaluate its centroid
            temp = np.where(y==self.class_list[i][0])[0]
            self.centroids[i,:] = np.mean(X[temp],axis=0)
            
        # Covariance matrices evaluation
        C_list = [np.zeros((X_tr_norm.shape[1],X_tr_norm.shape[1]))]*len(class_list)
        for i in range(len(class_list)):
            # points of class class_list[i]:
            temp = X[np.where(y==self.class_list[i][0])[0]]
            for datapoint in temp:
                aux = np.asmatrix(datapoint - centroids[i])
                C_list[i] += np.matmul(aux.T,aux)
            C_list[i] /= len(temp) # divide by number of samples
        
        if self.variant is None:
            self.C_inv = [inv(C) for C in C_list]
        
        elif self.variant==1: # diagonal
            self.C_inv = [inv(np.diag(np.diag(C))) for C in C_list]
        
        elif self.variant==2: # pooled
            self.C_inv = inv(self.__C_pool(X, y))
        
        elif self.variant==3: # Friedman regularization
            self.C_inv = [np.zeros((X.shape[1],X.shape[1]))]*len(self.class_list)
            N = len(y) # total number of samples
            for i in range(len(self.class_list)): # for each class
                Ni = len(X[np.where(y==self.class_list[i][0])[0]]) # number of samples of class 'i'
                self.C_inv[i] = inv(
                    ((1-self.Lambda)*Ni*C_list[i] + self.Lambda*N*self.__C_pool(X, y)) /
                    ((1-self.Lambda)*Ni           + self.Lambda*N) 
                )
            
    def __C_pool(self, X, y):
        C_poll = np.zeros((X.shape[1],X.shape[1]))
        for i in range(len(self.class_list)):
            # C_poll = sum (Ni/N) * Ci
            C_poll += ( len(X[np.where(y==self.class_list[i][0])[0]]) / len(y) ) * C_list[i]
        
        return C_poll
            
    def predict(self, X):
        N = len(X)
        y_pred = np.zeros(N)
        for i in range(N): # for each sample
            # find closest
            dist = np.inf
            closest = None
            for j in range(len(self.class_list)): # calculate the distance from each class
                temp = np.asmatrix(X[i] - self.centroids[j])
                # On variant 2 we have only one covariance matrix, in the other cases we have one covariance 
                # matrix for each class:
                C_inv = self.C_inv if variant==2 else self.C_inv[j]
                new_dist = np.matmul(np.matmul(temp,C_inv), temp.T)
                if new_dist < dist:
                    dist = new_dist
                    closest = self.class_list[j][0]
            
            y_pred[i] = closest
                        
        return y_pred

%%time

n_resamplings=10

variants = [1, 2, 3]
lambdas = np.linspace(0,1,num=12)[1:-1] # we crop 0 and 1 because:
                                        # lambda=0 => original QGC
                                        # lambda=1 => variant 2 of QGC
header = ["variant","$\lambda$", "$\mu$", "$\sigma$", "$f_o$"]
qgc_data = np.empty((2 + len(lambdas), len(header)), dtype=object)

for variant in variants:
    print("Started variant {} at {}".format(variant, datetime.datetime.now()))
    results = [0]*n_resamplings if variant!=3 else np.zeros((len(lambdas),n_resamplings))
    for i in range(n_resamplings):
        # Train/validation split
        X_train, X_test, y_train, y_test = train_test_split(X_tr.values, Y_tr.values, test_size=test_size)
        
        # scaling
        X_tr_norm, X_ts_norm = scale_feat(X_train, X_test, featIndex=numFeat_idx, scaleType="min-max")
        
        if variant!=3:
            # model fitting
            qgc = QuadraticGaussianClassifier(variant=variant)
            qgc.fit(X_tr_norm, y_train)
            
            # model evaluation
            y_pred = qgc.predict(X_ts_norm)
            results[i] = f1_score(y_test, y_pred, average='weighted')
            
        else:
            for j in range(len(lambdas)): # for each lambda
                # model fitting
                qgc = QuadraticGaussianClassifier(variant=variant, Lambda=lambdas[j])
                qgc.fit(X_tr_norm, y_train)
                
                # model evaluation
                y_pred = qgc.predict(X_ts_norm)
                results[j,i] = f1_score(y_test, y_pred, average='weighted')
        
    if variant!=3:
        qgc_data[variant-1,:] = np.asmatrix(
            [variant, np.nan, np.mean(results), np.std(results), f_o(results)]
        )

    else:
        var3_matrix = np.asmatrix([3]*len(lambdas)).T
        lambdas_matrix  = np.asmatrix(lambdas).T
        fo = [f_o(result) for result in results]
        fo = np.asmatrix(fo).T
        qgc_data[2:2+len(lambdas),:] = np.concatenate(
            (var3_matrix, lambdas_matrix, np.asmatrix(np.mean(results,axis=1)).T, 
             np.asmatrix(np.std(results,axis=1)).T, fo), axis=1
        )


df_qgc = pd.DataFrame(qgc_data, columns=header)
display(df_qgc)
is_over()

# lost OUTPUT:
'''
Started variant 1 at 2019-06-18 21:06:56.687818
Started variant 2 at 2019-06-18 21:17:35.235734
Started variant 3 at 2019-06-18 21:28:02.312084

CPU times: user 6h 20min 15s, sys: 4h 36min 1s, total: 10h 56min 17s
Wall time: 2h 7min 30s
'''

# saving results
#df_qgc.to_csv("./simulation_results/df_qgc.csv", sep='\t', index=False)

# loading results
df_qgc = pd.read_csv("./simulation_results/df_qgc.csv", sep='\t')
print("df_qgc:")
df_qgc

## 3.6 Decison trees

%%time
from xgboost import XGBClassifier
import datetime

cases = [
    {
    "learning_rate"     : learning_rate
    ,'n_estimators'     : n_estimators
    ,'max_depth'        : max_depth
    ,'tree_method'      : 'gpu_hist'
    ,'objective'        : 'multi:softmax'
    } 
    # hyperparameters possible values
    for learning_rate    in [1e-1, 0.3, 0.5, 1]
    for n_estimators     in [500, 1000]
    for max_depth        in [6, 12]
]

header = list(cases[0].keys())[:-2] + ["$\mu$", "$\sigma$", "$f_o$"] # no need for repeating
                                                                     # 'tree_method' and 'objective'

xgb_data = np.empty((len(cases), len(header)), dtype=object)
count=0
for case in cases:
    print("Starting instance {}/{} at {}".format(count+1, len(cases), datetime.datetime.now()))
    results = [0]*n_resamplings 
    for i in range(n_resamplings):
        # Train/test split
        X_train, X_test, y_train, y_test = train_test_split(X_tr.values, Y_tr.values, 
                                                            test_size=test_size)
        # Train/validation split
        X_train, X_validation, y_train, y_validation = train_test_split(X_train, y_train, 
                                                                        test_size=test_size)

        # no scaling is needed!!!
        
        # model fitting
        xgb = XGBClassifier(**case)
        xgb.fit(X_train, y_train, eval_set=[(X_validation,y_validation)]
                ,early_stopping_rounds=30, verbose=False)
                
        # model evaluation
        y_pred = xgb.predict(X_test)
        results[i] = f1_score(y_test, y_pred, average='weighted')
            

    xgb_data[count,:] = np.matrix(
        list(case.values())[:-2] + [np.mean(results), np.std(results), f_o(results)])
    count+=1
        
    
df_xgb = pd.DataFrame(xgb_data, columns=header)
display(df_xgb)
is_over()

print("Best hyper-parameters:")
df_xgb.sort_values("$f_o$", ascending=False).iloc[0,:].to_frame().transpose()

# saving results
#df_xgb.to_csv("./simulation_results/df_xgb.csv", sep='\t', index=False)

# loading results
df_xgb = pd.read_csv("./simulation_results/df_xgb.csv", sep='\t')
#df_xgb

## 3.7 Artificial Neural Network

%%time 
from keras.models import Sequential
from keras.layers import Dense
from keras.optimizers import Adam
from keras.utils import to_categorical
from keras import backend as K
from keras.callbacks import EarlyStopping

from matplotlib import pyplot

# the f1 function below was not developed by me, available at:
# https://stackoverflow.com/questions/43547402/how-to-calculate-f1-macro-in-keras
def f1(y_true, y_pred):
    def recall(y_true, y_pred):
        """Recall metric.
        Only computes a batch-wise average of recall.
        Computes the recall, a metric for multi-label classification of
        how many relevant items are selected.
        """
        true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
        possible_positives = K.sum(K.round(K.clip(y_true, 0, 1)))
        recall = true_positives / (possible_positives + K.epsilon())
        return recall

    def precision(y_true, y_pred):
        """Precision metric.
        Only computes a batch-wise average of precision.
        Computes the precision, a metric for multi-label classification of
        how many selected items are relevant.
        """
        true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
        predicted_positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
        precision = true_positives / (predicted_positives + K.epsilon())
        return precision
    precision = precision(y_true, y_pred)
    recall = recall(y_true, y_pred)
    return 2*((precision*recall)/(precision+recall+K.epsilon()))

Y_tr_categorical = to_categorical(Y_tr)[:, 1:] # change to categorical

cases = [
    {
        'lr'  : lr,
        'arq' : arq
    } 
    # hyperparameters possible values
    for arq in [
                 [32],[128],[512]         # MLP with one   layer
                ,[512,128], [128,32]      # //   //  two   layers
                ,[512,128,32], [128,32,8] # //   //  three  //
                ]
    for lr in [1e-3, 1e-2]
]

header = list(cases[0].keys()) + ["$\mu$", "$\sigma$", "$f_o$"] 

ann_data = np.empty((len(cases), len(header)), dtype=object)
count=0
for case in cases:
    print("Started case {}/{} at {}".format(count+1, len(cases),datetime.datetime.now()))

    # model building
    arq = case['arq']
    model = Sequential()
    model.add(Dense(units=arq[0], activation='relu', input_dim=X_tr.shape[1])) # first layer
    for layer in arq:
        model.add(Dense(units=layer, activation='relu'))
    model.add(Dense(units=5, activation='softmax')) # output layer
    
    
    model.compile(loss='categorical_crossentropy'
                 ,metrics=[f1]
                 ,optimizer=Adam(lr=case['lr'], amsgrad=True)
                 )

    results = [0]*n_resamplings
    for i in range(n_resamplings):
        # Train/validation split
        X_train, X_test, y_train, y_test = train_test_split(X_tr.values, Y_tr_categorical, test_size=test_size)

        # scaling
        X_tr_norm, X_ts_norm = scale_feat(X_train, X_test, featIndex=numFeat_idx, scaleType="min-max")

        # model fitting
        es = EarlyStopping(monitor='val_f1', mode='max', verbose=0, patience=5)
        history = model.fit(X_tr_norm, y_train,shuffle=True 
                            ,epochs=1_000
                            ,verbose=0
                            ,validation_split=0.2
                            ,callbacks=[es]
                            )
    
        # model evaluation
        y_pred = np.argmax(model.predict(X_ts_norm), axis=1)+1
        y_pred = to_categorical(y_pred)[:, 1:]
        results[i] = f1_score(y_test, y_pred, average='weighted')
                
    ann_data[count,:] = np.matrix(list(case.values()) + [np.mean(results), np.std(results), f_o(results)])
    count+=1
        
        
df_ann = pd.DataFrame(ann_data, columns=header)
display(df_ann)
is_over()

# saving results
# df_ann.to_csv("./simulation_results/df_ann.csv", sep='\t', index=False)

# loading results
df_ann = pd.read_csv("./simulation_results/df_ann.csv", sep='\t')
# df_ann

print("Best hyper-parameters:")
df_ann.sort_values("$f_o$", ascending=False).iloc[0,:].to_frame().transpose()
